fix(ui): make app info reactive in modals to resolve v? fallback
